+++ SuspensionAnalysis/src/SuspensionAnalysis.Core/NumericalMethods/DifferentialEquation/Newmark/INewmarkMethod.cs
@@ -1,0 +1,7 @@
+﻿namespace SuspensionAnalysis.Core.NumericalMethods.DifferentialEquation.Newmark
+{
+    /// <summary>
+    /// It is responsible to execute the Newmark numerical method to solve Differential Equation.
+    /// </summary>
+    public interface INewmarkMethod : IDifferentialEquationMethod { }
+}