+++ SuspensionAnalysis/src/SuspensionAnalysis.Core/NumericalMethods/DifferentialEquation/NewmarkBeta/INewmarkBetaMethod.cs
@@ -1,0 +1,7 @@
+﻿namespace SuspensionAnalysis.Core.NumericalMethods.DifferentialEquation.NewmarkBeta
+{
+    /// <summary>
+    /// It is responsible to execute the Newmark-Beta numerical method to solve Differential Equation.
+    /// </summary>
+    public interface INewmarkBetaMethod : IDifferentialEquationMethod { }
+}