+++ SuspensionAnalysis/src/SuspensionAnalysis.Core/Operations/RunAnalysis/Static/CircularProfile/RunCircularProfileStaticAnalysis.cs
@@ -4,12 +4,12 @@ using SuspensionAnalysis.Core.Mapper;
 using SuspensionAnalysis.Core.Operations.CalculateReactions;
 using DataContract = SuspensionAnalysis.DataContracts.Models.Profiles;
 
-namespace SuspensionAnalysis.Core.Operations.RunAnalysis.CircularProfile
+namespace SuspensionAnalysis.Core.Operations.RunAnalysis.Static.CircularProfile
 {
     /// <summary>
     /// It is responsible to run the analysis to suspension system considering circular profile.
     /// </summary>
-    public class RunCircularProfileAnalysis : RunAnalysis<DataContract.CircularProfile>, IRunCircularProfileAnalysis
+    public class RunCircularProfileStaticAnalysis : RunStaticAnalysis<DataContract.CircularProfile>, IRunCircularProfileStaticAnalysis
     {
         /// <summary>
         /// Class constructor.
@@ -18,7 +18,7 @@ namespace SuspensionAnalysis.Core.Operations.RunAnalysis.CircularProfile
         /// <param name="mechanicsOfMaterials"></param>
         /// <param name="geometricProperty"></param>
         /// <param name="mappingResolver"></param>
-        public RunCircularProfileAnalysis(
+        public RunCircularProfileStaticAnalysis(
             ICalculateReactions calculateReactions,
             ICircularProfileMechanicsOfMaterials mechanicsOfMaterials,
             ICircularProfileGeometricProperty geometricProperty,