+++ SuspensionAnalysis/src/SuspensionAnalysis.Core/Operations/RunAnalysis/Static/RectangularProfile/IRunRectangularProfileStaticAnalysis.cs
@@ -1,9 +1,9 @@
 ﻿using DataContract = SuspensionAnalysis.DataContracts.Models.Profiles;
 
-namespace SuspensionAnalysis.Core.Operations.RunAnalysis.RectangularProfile
+namespace SuspensionAnalysis.Core.Operations.RunAnalysis.Static.RectangularProfile
 {
     /// <summary>
     /// It is responsible to run the analysis to suspension system considering rectangular profile.
     /// </summary>
-    public interface IRunRectangularProfileAnalysis : IRunAnalysis<DataContract.RectangularProfile> { }
+    public interface IRunRectangularProfileStaticAnalysis : IRunStaticAnalysis<DataContract.RectangularProfile> { }
 }